Code review and arrangement

+++ Entities/Concrete/ExchhangeRate.cs
@@ -10,6 +10,6 @@ namespace Entities.Concrete
         public int ExchhangeRateId { get; set; }
         public decimal USD { get; set; }
         public decimal EURO { get; set; }
-        public int STERLIN { get; set; }
+        public decimal STERLIN { get; set; }
     }
 }
